test: refactor jwt test & fix test error
